[feat] #62 공통으로 사용되는 adapter common 파일로 이동
