+++ app/src/main/java/com/sopt/geonppang/presentation/bakeryList/BakeryListAdapter.kt
@@ -1,21 +1,15 @@
-package com.sopt.geonppang.presentation.search
+package com.sopt.geonppang.presentation.bakeryList
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.sopt.geonppang.databinding.ItemBakeryBinding
 import com.sopt.geonppang.domain.model.Bakery
-import com.sopt.geonppang.util.ItemDiffCallback
 
-class BakeryAdapter(
+class BakeryListAdapter(
     private val moveToDetail: (Int) -> Unit
-) : ListAdapter<Bakery, BakeryAdapter.BakeryViewHolder>(
-    ItemDiffCallback<Bakery>(
-        onItemsTheSame = { old, new -> old.bakeryId == new.bakeryId },
-        onContentsTheSame = { old, new -> old == new }
-    )
-) {
+) : RecyclerView.Adapter<BakeryListAdapter.BakeryViewHolder>() {
+    private val bakeryList: MutableList<Bakery> = mutableListOf()
 
     class BakeryViewHolder(
         private val binding: ItemBakeryBinding,
@@ -38,7 +32,15 @@ class BakeryAdapter(
         return BakeryViewHolder(binding)
     }
 
+    override fun getItemCount() = bakeryList.size
+
     override fun onBindViewHolder(holder: BakeryViewHolder, position: Int) {
-        holder.onBind(getItem(position), moveToDetail)
+        holder.onBind(bakeryList[position], moveToDetail)
+    }
+
+    fun setGoalList(bakeries: MutableList<Bakery>) {
+        bakeryList.clear()
+        bakeryList.addAll(bakeries)
+        notifyDataSetChanged()
     }
 }
