+++ app/src/main/java/com/sopt/geonppang/presentation/common/BakeryAdapter.kt
@@ -1,15 +1,21 @@
-package com.sopt.geonppang.presentation.bakeryList
+package com.sopt.geonppang.presentation.common
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
+import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.sopt.geonppang.databinding.ItemBakeryBinding
 import com.sopt.geonppang.domain.model.Bakery
+import com.sopt.geonppang.util.ItemDiffCallback
 
-class BakeryRecyclerViewAdapter(
+class BakeryAdapter(
     private val moveToDetail: (Int) -> Unit
-) : RecyclerView.Adapter<BakeryRecyclerViewAdapter.BakeryViewHolder>() {
-    private val bakeryList: MutableList<Bakery> = mutableListOf()
+) : ListAdapter<Bakery, BakeryAdapter.BakeryViewHolder>(
+    ItemDiffCallback<Bakery>(
+        onItemsTheSame = { old, new -> old.bakeryId == new.bakeryId },
+        onContentsTheSame = { old, new -> old == new }
+    )
+) {
 
     class BakeryViewHolder(
         private val binding: ItemBakeryBinding,
@@ -32,15 +38,7 @@ class BakeryRecyclerViewAdapter(
         return BakeryViewHolder(binding)
     }
 
-    override fun getItemCount() = bakeryList.size
-
     override fun onBindViewHolder(holder: BakeryViewHolder, position: Int) {
-        holder.onBind(bakeryList[position], moveToDetail)
-    }
-
-    fun setGoalList(bakeries: MutableList<Bakery>) {
-        bakeryList.clear()
-        bakeryList.addAll(bakeries)
-        notifyDataSetChanged()
+        holder.onBind(getItem(position), moveToDetail)
     }
 }
