+++ app/src/main/java/com/sopt/geonppang/presentation/mypage/MyStoreListActivity.kt
@@ -8,7 +8,7 @@ import androidx.lifecycle.lifecycleScope
 import com.sopt.geonppang.R
 import com.sopt.geonppang.databinding.ActivityMyStoreListBinding
 import com.sopt.geonppang.presentation.detail.DetailActivity
-import com.sopt.geonppang.presentation.search.BakeryAdapter
+import com.sopt.geonppang.presentation.common.BakeryAdapter
 import com.sopt.geonppang.util.UiState
 import com.sopt.geonppang.util.binding.BindingActivity
 import dagger.hilt.android.AndroidEntryPoint
